+++ geo-types/src/multi_line_string.rs
@@ -1,11 +1,16 @@
 use crate::{CoordinateType, LineString};
 use std::iter::FromIterator;
 
-/// A collection of [`LineString`s](line_string/struct.LineString.html).
+/// A collection of
+/// [`LineString`s](line_string/struct.LineString.html). The
+/// interior and the boundary are the union of the interior
+/// or the boundary of the constituent line strings.
 ///
-/// Can be created from a `Vec` of `LineString`s, or from an Iterator which yields `LineString`s.
+/// Can be created from a `Vec` of `LineString`s, or from an
+/// Iterator which yields `LineString`s.
 ///
-/// Iterating over this objects, yields the component `LineString`s.
+/// Iterating over this objects, yields the component
+/// `LineString`s.
 #[derive(Eq, PartialEq, Clone, Debug, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct MultiLineString<T>(pub Vec<LineString<T>>)