+++ geo/src/algorithm/closest_point.rs
@@ -1,3 +1,4 @@
+use crate::kernels::*;
 use crate::prelude::*;
 use crate::{Closest, Line, LineString, MultiLineString, MultiPoint, MultiPolygon, Point, Polygon};
 use num_traits::Float;
@@ -47,7 +48,7 @@ impl<F: Float> ClosestPoint<F> for Point<F> {
     }
 }
 
-impl<F: Float> ClosestPoint<F> for Line<F> {
+impl<F: Float + HasKernel> ClosestPoint<F> for Line<F> {
     fn closest_point(&self, p: &Point<F>) -> Closest<F> {
         let line_length = self.euclidean_length();
         if line_length == F::zero() {
@@ -77,7 +78,7 @@ impl<F: Float> ClosestPoint<F> for Line<F> {
         let y = direction_vector.y();
         let c = Point(self.start + (t * x, t * y).into());
 
-        if self.contains(p) {
+        if self.intersects(p) {
             Closest::Intersection(c)
         } else {
             Closest::SinglePoint(c)
@@ -106,20 +107,20 @@ where
     best
 }
 
-impl<F: Float> ClosestPoint<F> for LineString<F> {
+impl<F: Float + HasKernel> ClosestPoint<F> for LineString<F> {
     fn closest_point(&self, p: &Point<F>) -> Closest<F> {
         closest_of(self.lines(), *p)
     }
 }
 
-impl<F: Float> ClosestPoint<F> for Polygon<F> {
+impl<F: Float + HasKernel> ClosestPoint<F> for Polygon<F> {
     fn closest_point(&self, p: &Point<F>) -> Closest<F> {
         let prospectives = self.interiors().iter().chain(iter::once(self.exterior()));
         closest_of(prospectives, *p)
     }
 }
 
-impl<F: Float> ClosestPoint<F> for MultiPolygon<F> {
+impl<F: Float + HasKernel> ClosestPoint<F> for MultiPolygon<F> {
     fn closest_point(&self, p: &Point<F>) -> Closest<F> {
         closest_of(self.0.iter(), *p)
     }
@@ -131,7 +132,7 @@ impl<F: Float> ClosestPoint<F> for MultiPoint<F> {
     }
 }
 
-impl<F: Float> ClosestPoint<F> for MultiLineString<F> {
+impl<F: Float + HasKernel> ClosestPoint<F> for MultiLineString<F> {
     fn closest_point(&self, p: &Point<F>) -> Closest<F> {
         closest_of(self.0.iter(), *p)
     }