+++ geo/src/algorithm/contains/mod.rs
@@ -1,434 +1,48 @@
-use num_traits::Float;
-
-use crate::algorithm::intersects::Intersects;
-use crate::utils;
-use crate::{
-    Coordinate, CoordinateType, Geometry, GeometryCollection, Line, LineString, MultiLineString,
-    MultiPoint, MultiPolygon, Point, Polygon, Rect, Triangle,
-};
-
-///  Checks if the geometry A is completely inside the B geometry
+/// Checks if `rhs` is completely contained within `self`.
+/// More formally, the interior of `rhs` has non-empty
+/// (set-theoretic) intersection but neither the interior,
+/// nor the boundary of `rhs` intersects the exterior of
+/// `self`. In other words, the [DE-9IM] intersection matrix
+/// of `(rhs, self)` is `T*F**F***`.
+///
+/// [DE-9IM]: https://en.wikipedia.org/wiki/DE-9IM
+///
+/// # Examples
+///
+/// ```
+/// use geo::algorithm::contains::Contains;
+/// use geo::{line_string, point, Polygon};
+///
+/// let line_string = line_string![
+///     (x: 0., y: 0.),
+///     (x: 2., y: 0.),
+///     (x: 2., y: 2.),
+///     (x: 0., y: 2.),
+///     (x: 0., y: 0.),
+/// ];
+///
+/// let polygon = Polygon::new(line_string.clone(), vec![]);
+///
+/// // Point in Point
+/// assert!(point!(x: 2., y: 0.).contains(&point!(x: 2., y: 0.)));
+///
+/// // Point in Linestring
+/// assert!(line_string.contains(&point!(x: 2., y: 0.)));
+///
+/// // Point in Polygon
+/// assert!(polygon.contains(&point!(x: 1., y: 1.)));
+/// ```
 pub trait Contains<Rhs = Self> {
-    /// Checks if `rhs` is completely contained within `self`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use geo::algorithm::contains::Contains;
-    /// use geo::{line_string, point, Polygon};
-    ///
-    /// let line_string = line_string![
-    ///     (x: 0., y: 0.),
-    ///     (x: 2., y: 0.),
-    ///     (x: 2., y: 2.),
-    ///     (x: 0., y: 2.),
-    ///     (x: 0., y: 0.),
-    /// ];
-    ///
-    /// let polygon = Polygon::new(line_string.clone(), vec![]);
-    ///
-    /// // Point in Point
-    /// assert!(point!(x: 2., y: 0.).contains(&point!(x: 2., y: 0.)));
-    ///
-    /// // Point in Linestring
-    /// assert!(line_string.contains(&point!(x: 2., y: 0.)));
-    ///
-    /// // Point in Polygon
-    /// assert!(polygon.contains(&point!(x: 1., y: 1.)));
-    /// ```
     fn contains(&self, rhs: &Rhs) -> bool;
 }
 
-// ┌───────────────────────────┐
-// │ Implementations for Point │
-// └───────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for Point<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.contains(&Point(*coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for Point<T>
-where
-    T: Float,
-{
-    fn contains(&self, p: &Point<T>) -> bool {
-        ::geo_types::private_utils::point_contains_point(*self, *p)
-    }
-}
-
-// ┌────────────────────────────────┐
-// │ Implementations for MultiPoint │
-// └────────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for MultiPoint<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.0.iter().any(|point| point.contains(coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for MultiPoint<T>
-where
-    T: Float,
-{
-    fn contains(&self, point: &Point<T>) -> bool {
-        self.contains(&point.0)
-    }
-}
-
-// ┌──────────────────────────┐
-// │ Implementations for Line │
-// └──────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for Line<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.contains(&Point(*coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for Line<T>
-where
-    T: Float,
-{
-    fn contains(&self, p: &Point<T>) -> bool {
-        self.intersects(p)
-    }
-}
-
-impl<T> Contains<Line<T>> for Line<T>
-where
-    T: Float,
-{
-    fn contains(&self, line: &Line<T>) -> bool {
-        self.contains(&line.start_point()) && self.contains(&line.end_point())
-    }
-}
-
-impl<T> Contains<LineString<T>> for Line<T>
-where
-    T: Float,
-{
-    fn contains(&self, linestring: &LineString<T>) -> bool {
-        linestring.points_iter().all(|pt| self.contains(&pt))
-    }
-}
-
-// ┌────────────────────────────────┐
-// │ Implementations for LineString │
-// └────────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for LineString<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.contains(&Point(*coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for LineString<T>
-where
-    T: Float,
-{
-    fn contains(&self, p: &Point<T>) -> bool {
-        self.lines().any(|line| line.contains(p))
-    }
-}
-
-impl<T> Contains<Line<T>> for LineString<T>
-where
-    T: Float,
-{
-    fn contains(&self, line: &Line<T>) -> bool {
-        let (p0, p1) = line.points();
-        let mut look_for: Option<Point<T>> = None;
-        for segment in self.lines() {
-            if look_for.is_none() {
-                // If segment contains an endpoint of line, we mark the other endpoint as the
-                // one we are looking for.
-                if segment.contains(&p0) {
-                    look_for = Some(p1);
-                } else if segment.contains(&p1) {
-                    look_for = Some(p0);
-                }
-            }
-            if let Some(p) = look_for {
-                // If we are looking for an endpoint, we need to either find it, or show that we
-                // should continue to look for it
-                if segment.contains(&p) {
-                    // If the segment contains the endpoint we are looking for we are done
-                    return true;
-                } else if !line.contains(&segment.end_point()) {
-                    // If not, and the end of the segment is not on the line, we should stop
-                    // looking
-                    look_for = None
-                }
-            }
-        }
-        false
-    }
-}
-
-// ┌─────────────────────────────────────┐
-// │ Implementations for MultiLineString │
-// └─────────────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for MultiLineString<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.0.iter().any(|line_string| line_string.contains(coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for MultiLineString<T>
-where
-    T: Float,
-{
-    fn contains(&self, point: &Point<T>) -> bool {
-        self.contains(&point.0)
-    }
-}
-
-// ┌─────────────────────────────┐
-// │ Implementations for Polygon │
-// └─────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for Polygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        match utils::coord_pos_relative_to_line_string(*coord, &self.exterior()) {
-            utils::CoordPos::OnBoundary | utils::CoordPos::Outside => false,
-            _ => self.interiors().iter().all(|ls| {
-                utils::coord_pos_relative_to_line_string(*coord, ls) == utils::CoordPos::Outside
-            }),
-        }
-    }
-}
-
-impl<T> Contains<Point<T>> for Polygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, p: &Point<T>) -> bool {
-        self.contains(&p.0)
-    }
-}
-
-impl<T> Contains<Line<T>> for Polygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, line: &Line<T>) -> bool {
-        // both endpoints are contained in the polygon and the line
-        // does NOT intersect the exterior or any of the interior boundaries
-        self.contains(&line.start_point())
-            && self.contains(&line.end_point())
-            && !self.exterior().intersects(line)
-            && !self.interiors().iter().any(|inner| inner.intersects(line))
-    }
-}
-
-impl<T> Contains<Polygon<T>> for Polygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, poly: &Polygon<T>) -> bool {
-        // decompose poly's exterior ring into Lines, and check each for containment
-        poly.exterior().lines().all(|line| self.contains(&line))
-    }
-}
-
-impl<T> Contains<LineString<T>> for Polygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, linestring: &LineString<T>) -> bool {
-        // All LineString points must be inside the Polygon
-        if linestring.points_iter().all(|point| self.contains(&point)) {
-            // The Polygon interior is allowed to intersect with the LineString
-            // but the Polygon's rings are not
-            !self
-                .interiors()
-                .iter()
-                .any(|ring| ring.intersects(linestring))
-        } else {
-            false
-        }
-    }
-}
-
-// ┌──────────────────────────────────┐
-// │ Implementations for MultiPolygon │
-// └──────────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for MultiPolygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.0.iter().any(|poly| poly.contains(coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for MultiPolygon<T>
-where
-    T: Float,
-{
-    fn contains(&self, p: &Point<T>) -> bool {
-        self.contains(&p.0)
-    }
-}
-
-// ┌──────────────────────────┐
-// │ Implementations for Rect │
-// └──────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for Rect<T>
-where
-    T: CoordinateType,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        coord.x >= self.min().x
-            && coord.x <= self.max().x
-            && coord.y >= self.min().y
-            && coord.y <= self.max().y
-    }
-}
-
-impl<T> Contains<Point<T>> for Rect<T>
-where
-    T: CoordinateType,
-{
-    fn contains(&self, p: &Point<T>) -> bool {
-        self.contains(&p.0)
-    }
-}
-
-impl<T> Contains<Rect<T>> for Rect<T>
-where
-    T: CoordinateType,
-{
-    fn contains(&self, bounding_rect: &Rect<T>) -> bool {
-        // All points of LineString must be in the polygon ?
-        self.min().x <= bounding_rect.min().x
-            && self.max().x >= bounding_rect.max().x
-            && self.min().y <= bounding_rect.min().y
-            && self.max().y >= bounding_rect.max().y
-    }
-}
-
-// ┌──────────────────────────────┐
-// │ Implementations for Triangle │
-// └──────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for Triangle<T>
-where
-    T: CoordinateType,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        let s = self.0.y * self.2.x - self.0.x * self.2.y
-            + (self.2.y - self.0.y) * coord.x
-            + (self.0.x - self.2.x) * coord.y;
-        let t = self.0.x * self.1.y - self.0.y * self.1.x
-            + (self.0.y - self.1.y) * coord.x
-            + (self.1.x - self.0.x) * coord.y;
-
-        if (s < T::zero()) != (t < T::zero()) {
-            return false;
-        }
-
-        let a = (T::zero() - self.1.y) * self.2.x
-            + self.0.y * (self.2.x - self.1.x)
-            + self.0.x * (self.1.y - self.2.y)
-            + self.1.x * self.2.y;
-
-        if a == T::zero() {
-            false
-        } else if a < T::zero() {
-            s <= T::zero() && s + t >= a
-        } else {
-            s >= T::zero() && s + t <= a
-        }
-    }
-}
-
-impl<T> Contains<Point<T>> for Triangle<T>
-where
-    T: CoordinateType,
-{
-    fn contains(&self, point: &Point<T>) -> bool {
-        self.contains(&point.0)
-    }
-}
-
-// ┌──────────────────────────────┐
-// │ Implementations for Geometry │
-// └──────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for Geometry<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        match self {
-            Geometry::Point(g) => g.contains(coord),
-            Geometry::Line(g) => g.contains(coord),
-            Geometry::LineString(g) => g.contains(coord),
-            Geometry::Polygon(g) => g.contains(coord),
-            Geometry::MultiPoint(g) => g.contains(coord),
-            Geometry::MultiLineString(g) => g.contains(coord),
-            Geometry::MultiPolygon(g) => g.contains(coord),
-            Geometry::GeometryCollection(g) => g.contains(coord),
-            Geometry::Rect(g) => g.contains(coord),
-            Geometry::Triangle(g) => g.contains(coord),
-        }
-    }
-}
-
-impl<T> Contains<Point<T>> for Geometry<T>
-where
-    T: Float,
-{
-    fn contains(&self, point: &Point<T>) -> bool {
-        self.contains(&point.0)
-    }
-}
-
-// ┌────────────────────────────────────────┐
-// │ Implementations for GeometryCollection │
-// └────────────────────────────────────────┘
-
-impl<T> Contains<Coordinate<T>> for GeometryCollection<T>
-where
-    T: Float,
-{
-    fn contains(&self, coord: &Coordinate<T>) -> bool {
-        self.0.iter().any(|geometry| geometry.contains(coord))
-    }
-}
-
-impl<T> Contains<Point<T>> for GeometryCollection<T>
-where
-    T: Float,
-{
-    fn contains(&self, point: &Point<T>) -> bool {
-        self.contains(&point.0)
-    }
-}
+mod geometry;
+mod line;
+mod line_string;
+mod point;
+mod polygon;
+mod rect;
+mod triangle;
 
 // ┌───────┐
 // │ Tests │
@@ -524,7 +138,11 @@ mod test {
     #[test]
     fn linestring_point_is_vertex_test() {
         let linestring = LineString::from(vec![(0., 0.), (2., 0.), (2., 2.)]);
-        assert!(linestring.contains(&Point::new(2., 2.)));
+        // Note: the end points of a linestring are not
+        // considered to be "contained"
+        assert!(linestring.contains(&Point::new(2., 0.)));
+        assert!(!linestring.contains(&Point::new(0., 0.)));
+        assert!(!linestring.contains(&Point::new(2., 2.)));
     }
     #[test]
     fn linestring_test() {
@@ -709,30 +327,62 @@ mod test {
     }
     #[test]
     fn linestring_in_line_test() {
-        let line = Line::from([(0., 1.), (3., 4.)]);
+        let line = Line::from([(0, 10), (30, 40)]);
         // linestring0 in line
-        let linestring0 = LineString::from(vec![(0.1, 1.1), (1., 2.), (1.5, 2.5)]);
+        let linestring0 = LineString::from(vec![(01, 11), (10, 20), (15, 25)]);
         // linestring1 starts and ends in line, but wanders in the middle
-        let linestring1 = LineString::from(vec![(0.1, 1.1), (2., 2.), (1.5, 2.5)]);
+        let linestring1 = LineString::from(vec![(01, 11), (20, 20), (15, 25)]);
         // linestring2 is co-linear, but extends beyond line
-        let linestring2 = LineString::from(vec![(0.1, 1.1), (1., 2.), (4., 5.)]);
+        let linestring2 = LineString::from(vec![(01, 11), (10, 20), (40, 50)]);
         // no part of linestring3 is contained in line
-        let linestring3 = LineString::from(vec![(1.1, 1.1), (2., 2.), (2.5, 2.5)]);
+        let linestring3 = LineString::from(vec![(11, 11), (20, 20), (25, 25)]);
+        // a linestring with singleton interior on the boundary of the line
+        let linestring4 = LineString::from(vec![(0, 10), (0, 10), (0, 10)]);
+        // a linestring with singleton interior that is contained in the line
+        let linestring5 = LineString::from(vec![(1, 11), (1, 11), (1, 11)]);
         assert!(line.contains(&linestring0));
         assert!(!line.contains(&linestring1));
         assert!(!line.contains(&linestring2));
         assert!(!line.contains(&linestring3));
+        assert!(!line.contains(&linestring4));
+        assert!(line.contains(&linestring5));
     }
     #[test]
     fn line_in_polygon_test() {
         let c = |x, y| Coordinate { x, y };
-        let line = Line::new(c(0., 1.), c(3., 4.));
-        let linestring0 = line_string![c(-1., 0.), c(5., 0.), c(5., 5.), c(0., 5.), c(-1., 0.)];
+        let line = Line::new(c(0, 10), c(30, 40));
+        let linestring0 = line_string![c(-10, 0), c(50, 0), c(50, 50), c(0, 50), c(-10, 0)];
         let poly0 = Polygon::new(linestring0, Vec::new());
-        let linestring1 = line_string![c(0., 0.), c(0., 2.), c(2., 2.), c(2., 0.), c(0., 0.)];
+        let linestring1 = line_string![c(0, 0), c(0, 20), c(20, 20), c(20, 0), c(0, 0)];
         let poly1 = Polygon::new(linestring1, Vec::new());
         assert!(poly0.contains(&line));
         assert!(!poly1.contains(&line));
+    }
+    #[test]
+    #[ignore]
+    fn line_in_polygon_edgecases_test() {
+        // Some DE-9IM edge cases for checking line is
+        // inside polygon The end points of the line can be
+        // on the boundary of the polygon but we don't allow
+        // that yet.
+        let c = |x, y| Coordinate { x, y };
+        // A non-convex polygon
+        let linestring0 = line_string![c(0, 0), c(1, 1), c(1, -1), c(-1, -1), c(-1, 1)];
+        let poly = Polygon::new(linestring0, Vec::new());
+
+        assert!(poly.contains(&Line::new(c(0, 0), c(1, -1))));
+        assert!(poly.contains(&Line::new(c(-1, 1), c(1, -1))));
+        assert!(!poly.contains(&Line::new(c(-1, 1), c(1, 1))));
+    }
+    #[test]
+    fn line_in_linestring_edgecases() {
+        let c = |x, y| Coordinate { x, y };
+        use crate::line_string;
+        let mut ls = line_string![c(0, 0), c(1, 0), c(0, 1), c(-1, 0)];
+        assert!(!ls.contains(&Line::from([(0, 0), (0, 0)])));
+        ls.close();
+        assert!(ls.contains(&Line::from([(0, 0), (0, 0)])));
+        assert!(ls.contains(&Line::from([(-1, 0), (1, 0)])));
     }
     #[test]
     fn line_in_linestring_test() {
@@ -775,17 +425,17 @@ mod test {
     }
 
     #[test]
-    fn triangle_contains_point_on_edge() {
+    fn triangle_not_contains_point_on_edge() {
         let t = Triangle::from([(0.0, 0.0), (2.0, 0.0), (2.0, 2.0)]);
         let p = Point::new(1.0, 0.0);
-        assert!(t.contains(&p));
+        assert!(!t.contains(&p));
     }
 
     #[test]
-    fn triangle_contains_point_on_vertex() {
+    fn triangle_not_contains_point_on_vertex() {
         let t = Triangle::from([(0.0, 0.0), (2.0, 0.0), (2.0, 2.0)]);
         let p = Point::new(2.0, 0.0);
-        assert!(t.contains(&p));
+        assert!(!t.contains(&p));
     }
 
     #[test]