+++ geo/src/algorithm/euclidean_distance.rs
@@ -3,7 +3,7 @@ use crate::algorithm::euclidean_length::EuclideanLength;
 use crate::algorithm::intersects::Intersects;
 use crate::algorithm::polygon_distance_fast_path::*;
 use crate::kernels::*;
-use crate::utils::{coord_pos_relative_to_line_string, CoordPos};
+use crate::utils::{coord_pos_relative_to_ring, CoordPos};
 use crate::{
     Line, LineString, MultiLineString, MultiPoint, MultiPolygon, Point, Polygon, Triangle,
 };
@@ -132,7 +132,7 @@ where
 
 impl<T> EuclideanDistance<T, Polygon<T>> for Point<T>
 where
-    T: Float,
+    T: Float + HasKernel,
 {
     /// Minimum distance from a Point to a Polygon
     fn euclidean_distance(&self, polygon: &Polygon<T>) -> T {
@@ -165,7 +165,7 @@ where
 
 impl<T> EuclideanDistance<T, Point<T>> for Polygon<T>
 where
-    T: Float,
+    T: Float + HasKernel,
 {
     /// Minimum distance from a Polygon to a Point
     fn euclidean_distance(&self, point: &Point<T>) -> T {
@@ -175,7 +175,7 @@ where
 
 impl<T> EuclideanDistance<T, MultiPolygon<T>> for Point<T>
 where
-    T: Float,
+    T: Float + HasKernel,
 {
     /// Minimum distance from a Point to a MultiPolygon
     fn euclidean_distance(&self, mpolygon: &MultiPolygon<T>) -> T {
@@ -189,7 +189,7 @@ where
 
 impl<T> EuclideanDistance<T, Point<T>> for MultiPolygon<T>
 where
-    T: Float,
+    T: Float + HasKernel,
 {
     /// Minimum distance from a MultiPolygon to a Point
     fn euclidean_distance(&self, point: &Point<T>) -> T {
@@ -263,7 +263,7 @@ where
 /// LineString-LineString distance
 impl<T> EuclideanDistance<T, LineString<T>> for LineString<T>
 where
-    T: Float + Signed + RTreeNum,
+    T: Float + HasKernel + Signed + RTreeNum,
 {
     fn euclidean_distance(&self, other: &LineString<T>) -> T {
         if self.intersects(other) {
@@ -280,9 +280,9 @@ where
 /// contain a point from the candidate Polygon's outer shell in their simple representations
 fn ring_contains_point<T>(poly: &Polygon<T>, p: Point<T>) -> bool
 where
-    T: Float,
+    T: HasKernel,
 {
-    match coord_pos_relative_to_line_string(p.0, &poly.exterior()) {
+    match coord_pos_relative_to_ring(p.0, &poly.exterior()) {
         CoordPos::Inside => true,
         CoordPos::OnBoundary | CoordPos::Outside => false,
     }
@@ -291,7 +291,7 @@ where
 /// LineString to Line
 impl<T> EuclideanDistance<T, Line<T>> for LineString<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &Line<T>) -> T {
         self.lines().fold(Bounded::max_value(), |acc, line| {
@@ -303,7 +303,7 @@ where
 /// Line to LineString
 impl<T> EuclideanDistance<T, LineString<T>> for Line<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &LineString<T>) -> T {
         other.euclidean_distance(self)
@@ -313,7 +313,7 @@ where
 /// LineString to Polygon
 impl<T> EuclideanDistance<T, Polygon<T>> for LineString<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &Polygon<T>) -> T {
         if self.intersects(other) || other.contains(self) {
@@ -334,7 +334,7 @@ where
 /// Polygon to LineString distance
 impl<T> EuclideanDistance<T, LineString<T>> for Polygon<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &LineString<T>) -> T {
         other.euclidean_distance(self)
@@ -344,7 +344,7 @@ where
 /// Line to MultiPolygon distance
 impl<T> EuclideanDistance<T, MultiPolygon<T>> for Line<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, mpolygon: &MultiPolygon<T>) -> T {
         mpolygon
@@ -358,7 +358,7 @@ where
 /// MultiPolygon to Line distance
 impl<T> EuclideanDistance<T, Line<T>> for MultiPolygon<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &Line<T>) -> T {
         other.euclidean_distance(self)
@@ -368,7 +368,7 @@ where
 /// Line to Line distance
 impl<T> EuclideanDistance<T, Line<T>> for Line<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &Line<T>) -> T {
         if self.intersects(other) || self.contains(other) {
@@ -386,7 +386,7 @@ where
 // Line to Polygon distance
 impl<T> EuclideanDistance<T, Polygon<T>> for Line<T>
 where
-    T: Float + Signed + RTreeNum + FloatConst,
+    T: Float + Signed + RTreeNum + FloatConst + HasKernel,
 {
     fn euclidean_distance(&self, other: &Polygon<T>) -> T {
         if other.contains(self) || self.intersects(other) {
@@ -420,7 +420,7 @@ where
 // Polygon to Line distance
 impl<T> EuclideanDistance<T, Line<T>> for Polygon<T>
 where
-    T: Float + FloatConst + Signed + RTreeNum,
+    T: Float + FloatConst + Signed + RTreeNum + HasKernel,
 {
     fn euclidean_distance(&self, other: &Line<T>) -> T {
         other.euclidean_distance(self)
@@ -468,7 +468,7 @@ where
 
 impl<T> EuclideanDistance<T, Point<T>> for Triangle<T>
 where
-    T: Float,
+    T: Float + HasKernel,
 {
     fn euclidean_distance(&self, point: &Point<T>) -> T {
         if self.contains(point) {
@@ -659,6 +659,7 @@ mod test {
         // A point inside the cutout triangle
         let p = Point::new(3.5, 2.5);
         let dist = p.euclidean_distance(&poly);
+
         // 0.41036467732879783 <-- Shapely
         assert_relative_eq!(dist, 0.41036467732879767);
     }