+++ geo/src/algorithm/kernels/robust.rs
@@ -1,6 +1,5 @@
 use super::{Kernel, Orientation};
 use crate::Coordinate;
-use std::marker::PhantomData;
 
 /// Robust kernel that uses [fast robust
 /// predicates](//www.cs.cmu.edu/~quake/robust.html) to
@@ -8,17 +7,11 @@ use std::marker::PhantomData;
 /// used with types that can _always_ be casted to `f64`
 /// _without loss in precision_.
 #[derive(Default)]
-pub struct RobustKernel<T>(PhantomData<T>);
+pub struct RobustKernel;
 
 use num_traits::{Float, NumCast};
-impl<T: Float> Kernel for RobustKernel<T> {
-    type Scalar = T;
-
-    fn orient2d(
-        p: Coordinate<Self::Scalar>,
-        q: Coordinate<Self::Scalar>,
-        r: Coordinate<Self::Scalar>,
-    ) -> Orientation {
+impl<T: Float> Kernel<T> for RobustKernel {
+    fn orient2d(p: Coordinate<T>, q: Coordinate<T>, r: Coordinate<T>) -> Orientation {
         use robust::{orient2d, Coord};
 
         let orientation = orient2d(